refactor(user): reorganize dtos into domain-specific directories
- Move support-related dtos to user/suporter directory
- Move auth-related dtos to user/auth directory
- Move artist-request dto to user/artist-request directory
- Fix indentation and formatting in several use case files
