+++ services/user-admin-service/src/application/usecases/user/user-intraction/GetUserWithIdUserUseCase.ts
@@ -1,8 +1,8 @@
 import { BadRequestError, NotFoundError } from "art-chain-shared";
 import { AUTH_MESSAGES } from "../../../../constants/authMessages";
-import { USER_MESSAGES } from './../../../../constants/userMessages';
+import { USER_MESSAGES } from "./../../../../constants/userMessages";
 import { IUserRepository } from "../../../../domain/repositories/user/IUserRepository";
-import { GetUserProfileWithIdDto } from "../../../../domain/dtos/user/GetUserProfileWithIdDto";
+import { GetUserProfileWithIdDto } from "../../../../domain/dtos/user/suporter/GetUserProfileWithIdRequestDto";
 import { ISupporterRepository } from "../../../../domain/repositories/user/ISupporterRepository";
 
 export class GetUserWithIdUserUseCase {
@@ -15,7 +15,7 @@ export class GetUserWithIdUserUseCase {
     const { userId, currentUserId } = data;
 
     if (!userId) {
-        throw new BadRequestError(USER_MESSAGES.USER_ID_REQUIRED)
+      throw new BadRequestError(USER_MESSAGES.USER_ID_REQUIRED);
     }
 
     let currentUser = null;
